Re-index mesh so each vertex only has one normal.
By default, the OBJLoader removes the indexing inherent in the
.obj file (see https://github.com/mrdoob/three.js/issues/11776).
This means that it is a "triangle soup", where each face is three
distinct triangles. When calculating vertex normals, each vertex
will thus just have the same normal as the face, so scaling along
the normals will explode the mesh! See code comments for the fix.